+++ yggdrasil_decision_forests/port/python/ydf/model/jax_model_test.py
@@ -55,20 +55,28 @@ def create_dataset(
       "i1": np.random.randint(100, size=n),
       "i2": np.random.randint(100, size=n),
       "c1": np.random.choice(["x", "y", "z"], size=n, p=[0.6, 0.3, 0.1]),
+      "c2": np.random.choice([b"x", b"y", b"z"], size=n, p=[0.6, 0.3, 0.1]),
+      "c3": np.random.choice([1, 2, 3], size=n, p=[0.6, 0.3, 0.1]),
+      "c4": np.random.randint(2, size=n).astype(np.bool_),
       "b1": np.random.randint(2, size=n).astype(np.bool_),
       "b2": np.random.randint(2, size=n).astype(np.bool_),
       # Cat-set features
       "cs1": [[], ["a", "b", "c"], ["b", "c"], ["a"]] * (n // 4),
+      "cs2": [[], [b"a", b"b", b"c"], [b"b", b"c"], [b"a"]] * (n // 4),
+      "cs3": [[], [1, 2, 3], [2, 3], [1]] * (n // 4),
+      "cs4": [[], [False, True, False], [True, False], [False]] * (n // 4),
       # Multi-dim features
       "multi_f1": np.random.random(size=(n, 5)),
       "multi_f2": np.random.random(size=(n, 5)),
       "multi_i1": np.random.randint(100, size=(n, 5)),
       "multi_c1": np.random.choice(["x", "y", "z"], size=(n, 5)),
+      "multi_c2": np.random.choice([b"x", b"y", b"z"], size=(n, 5)),
       "multi_b1": np.random.randint(2, size=(n, 5)).astype(np.bool_),
       # Labels
-      "label_class_binary": np.random.choice([False, True], size=n),
-      "label_class_multi": np.random.choice(["l1", "l2", "l3"], size=n),
-      "label_regress": np.random.random(size=n),
+      "label_class_binary1": np.random.choice([False, True], size=n),
+      "label_class_multi1": np.random.choice(["l1", "l2", "l3"], size=n),
+      "label_class_multi2": np.random.choice([b"l1", b"l2", b"l3"], size=n),
+      "label_regress1": np.random.random(size=n),
   }
   return {k: data[k] for k in columns}
 
@@ -215,9 +223,9 @@ def check_toy_model(test_self, model):
 
 
 def create_toy_model(test_self):
-  columns = ["f1", "c1", "f2", "label_regress"]
+  columns = ["f1", "c1", "f2", "c2", "label_regress1"]
   model = specialized_learners.GradientBoostedTreesLearner(
-      label="label_regress",
+      label="label_regress1",
       task=generic_learner.Task.REGRESSION,
       num_trees=1,
   ).train(create_dataset(columns))
@@ -386,11 +394,31 @@ class JaxModelTest(parameterized.TestCase):
     )
 
   def test_feature_encoder_on_model(self):
-    columns = ["f1", "i1", "c1", "b1", "cs1", "label_class_binary"]
+    columns = [
+        "f1",
+        "i1",
+        "c1",
+        "c2",
+        "c3",
+        "c4",
+        "b1",
+        "cs1",
+        "cs2",
+        "cs3",
+        "cs4",
+        "label_class_binary1",
+    ]
     model = specialized_learners.RandomForestLearner(
-        label="label_class_binary",
+        label="label_class_binary1",
         num_trees=2,
-        features=[("cs1", dataspec_lib.Semantic.CATEGORICAL_SET)],
+        features=[
+            ("c3", dataspec_lib.Semantic.CATEGORICAL),
+            ("c4", dataspec_lib.Semantic.CATEGORICAL),
+            ("cs1", dataspec_lib.Semantic.CATEGORICAL_SET),
+            ("cs2", dataspec_lib.Semantic.CATEGORICAL_SET),
+            ("cs3", dataspec_lib.Semantic.CATEGORICAL_SET),
+            ("cs4", dataspec_lib.Semantic.CATEGORICAL_SET),
+        ],
         include_all_columns=True,
     ).train(create_dataset(columns))
     feature_encoder = to_jax.FeatureEncoder.build(
@@ -400,29 +428,81 @@ class JaxModelTest(parameterized.TestCase):
         feature_encoder.categorical,
         {
             "c1": {"<OOD>": 0, "x": 1, "y": 2, "z": 3},
+            "c2": {"<OOD>": 0, "x": 1, "y": 2, "z": 3},
+            "c3": {"<OOD>": 0, "1": 1, "2": 2, "3": 3},
+            "c4": {"<OOD>": 0, "false": 1, "true": 2},
             "cs1": {"<OOD>": 0, "a": 1, "b": 2, "c": 3},
+            "cs2": {"<OOD>": 0, "a": 1, "b": 2, "c": 3},
+            "cs3": {"<OOD>": 0, "1": 1, "2": 2, "3": 3},
+            "cs4": {"<OOD>": 0, "False": 1, "True": 2},
         },
     )
 
-    encoded_features = feature_encoder.encode(
-        {"f1": [1, 2, 3], "c1": ["x", "y", "other"]}
+    feature_values_py = {
+        "f1": [1, 2, 3],
+        "c1": ["x", "y", "other"],
+        "c2": [b"x", b"y", b"other"],
+        "c3": [1, 2, 4],
+        "c4": [False, True],
+    }
+    encoded_features_py = feature_encoder.encode(feature_values_py)
+    np.testing.assert_array_equal(
+        encoded_features_py["f1"], jnp.asarray([1, 2, 3])
     )
     np.testing.assert_array_equal(
-        encoded_features["f1"], jnp.asarray([1, 2, 3])
+        encoded_features_py["c1"], jnp.asarray([1, 2, 0])
     )
     np.testing.assert_array_equal(
-        encoded_features["c1"], jnp.asarray([1, 2, 0])
+        encoded_features_py["c2"], jnp.asarray([1, 2, 0])
+    )
+    np.testing.assert_array_equal(
+        encoded_features_py["c3"], jnp.asarray([1, 2, 0])
+    )
+    np.testing.assert_array_equal(
+        encoded_features_py["c4"], jnp.asarray([1, 2])
     )
 
+    feature_values_np = {k: np.asarray(v) for k, v in feature_values_py.items()}
+    encoded_features_np = feature_encoder.encode(feature_values_np)
+    for k in encoded_features_np:
+      np.testing.assert_array_equal(
+          encoded_features_np[k], encoded_features_py[k]
+      )
+
   def test_feature_encoder_categorical_is_empty(self):
-    columns = ["f1", "i1", "label_class_binary"]
+    columns = ["f1", "i1", "label_class_binary1"]
     model = specialized_learners.RandomForestLearner(
-        label="label_class_binary", num_trees=2
+        label="label_class_binary1", num_trees=2
     ).train(create_dataset(columns))
     feature_encoder = to_jax.FeatureEncoder.build(
         model.input_features(), model.data_spec()
     )
     self.assertEmpty(feature_encoder.categorical)
+
+  def test_feature_encoder_unexpected_type(self):
+    columns = ["c1", "label_class_binary1"]
+    model = specialized_learners.RandomForestLearner(
+        label="label_class_binary1", num_trees=2
+    ).train(create_dataset(columns))
+    feature_encoder = to_jax.FeatureEncoder.build(
+        model.input_features(), model.data_spec()
+    )
+    encoded_features = feature_encoder.encode({"f1": [1.0, 2.0, 3.0]})
+    np.testing.assert_array_equal(
+        encoded_features["f1"], jnp.asarray([1.0, 2.0, 3.0])
+    )
+    with self.assertRaisesRegex(
+        ValueError, "Unexpected categorical value type"
+    ):
+      feature_encoder.encode({"c1": [1.0, 2.0, 3.0]})
+    with self.assertRaisesRegex(
+        ValueError, "Unexpected categorical value type"
+    ):
+      feature_encoder.encode({"c1": ["x", "y", None]})
+    with self.assertRaisesRegex(
+        ValueError, "Unexpected categorical value type"
+    ):
+      feature_encoder.encode({"c1": [["x", "y"], ["z", "w"]]})
 
 
 class InternalFeatureSpecTest(parameterized.TestCase):
@@ -446,8 +526,14 @@ class InternalFeatureSpecTest(parameterized.TestCase):
             generic_model.InputFeature(
                 "c2", dataspec_lib.Semantic.CATEGORICAL, 5
             ),
-            generic_model.InputFeature("b1", dataspec_lib.Semantic.BOOLEAN, 6),
-            generic_model.InputFeature("b2", dataspec_lib.Semantic.BOOLEAN, 7),
+            generic_model.InputFeature(
+                "c3", dataspec_lib.Semantic.CATEGORICAL, 6
+            ),
+            generic_model.InputFeature(
+                "c4", dataspec_lib.Semantic.CATEGORICAL, 7
+            ),
+            generic_model.InputFeature("b1", dataspec_lib.Semantic.BOOLEAN, 8),
+            generic_model.InputFeature("b2", dataspec_lib.Semantic.BOOLEAN, 9),
         ],
         ds_pb.DataSpecification(
             created_num_rows=3,
@@ -476,6 +562,14 @@ class InternalFeatureSpecTest(parameterized.TestCase):
                 ),
                 ds_pb.Column(
                     name="c2",
+                    type=ds_pb.ColumnType.CATEGORICAL,
+                ),
+                ds_pb.Column(
+                    name="c3",
+                    type=ds_pb.ColumnType.CATEGORICAL,
+                ),
+                ds_pb.Column(
+                    name="c4",
                     type=ds_pb.ColumnType.CATEGORICAL,
                 ),
                 ds_pb.Column(
@@ -511,6 +605,8 @@ class InternalFeatureSpecTest(parameterized.TestCase):
         [
             InternalFeatureItem(name="c1", dim=1),
             InternalFeatureItem(name="c2", dim=1),
+            InternalFeatureItem(name="c3", dim=1),
+            InternalFeatureItem(name="c4", dim=1),
         ],
     )
     self.assertEqual(
@@ -521,11 +617,13 @@ class InternalFeatureSpecTest(parameterized.TestCase):
         ],
     )
     self.assertEqual(self.basic_mapping.inv_numerical, {0: 2, 1: 3, 2: 0, 3: 1})
-    self.assertEqual(self.basic_mapping.inv_categorical, {4: 0, 5: 1})
-    self.assertEqual(self.basic_mapping.inv_boolean, {6: 0, 7: 1})
+    self.assertEqual(
+        self.basic_mapping.inv_categorical, {4: 0, 5: 1, 6: 2, 7: 3}
+    )
+    self.assertEqual(self.basic_mapping.inv_boolean, {8: 0, 9: 1})
     self.assertEqual(
         self.basic_mapping.feature_names,
-        {"multidim_n3", "n1", "n2", "c1", "c2", "b1", "b2"},
+        {"multidim_n3", "n1", "n2", "c1", "c2", "c3", "c4", "b1", "b2"},
     )
 
   def test_non_supported_type(self):
@@ -554,9 +652,11 @@ class InternalFeatureSpecTest(parameterized.TestCase):
       self.basic_mapping.convert_features({
           "n1": jnp.array([1, 2]),
           "n2": jnp.array([3, 4]),
-          "multidim_n3": jnp.array([[9, 10], [11, 12]]),
+          "multidim_n3": jnp.array([[13, 14], [15, 16]]),
           "c1": jnp.array([5, 6]),
           "c2": jnp.array([7, 8]),
+          "c3": jnp.array([9, 10]),
+          "c4": jnp.array([11, 12]),
           "b1": jnp.array([True, False]),
           "b2": jnp.array([False, True]),
           "other": jnp.array([1, 2]),
@@ -569,9 +669,11 @@ class InternalFeatureSpecTest(parameterized.TestCase):
           "n2": jnp.array([3, 4]),
           # multidim_n3 is a 3-dimensional feature, but 1-dimentional values are
           # fed (taking into account the batch size).
-          "multidim_n3": jnp.array([9, 10]),
+          "multidim_n3": jnp.array([13, 14]),
           "c1": jnp.array([5, 6]),
           "c2": jnp.array([7, 8]),
+          "c3": jnp.array([9, 10]),
+          "c4": jnp.array([11, 12]),
           "b1": jnp.array([True, False]),
           "b2": jnp.array([False, True]),
       })
@@ -583,9 +685,11 @@ class InternalFeatureSpecTest(parameterized.TestCase):
           "n2": jnp.array([[9, 10], [11, 12]]),
           # multidim_n3 is a 3-dimensional feature, but 2-dimentional values are
           # fed (taking into account the batch size).
-          "multidim_n3": jnp.array([[9, 10], [11, 12]]),
+          "multidim_n3": jnp.array([[13, 14], [15, 16]]),
           "c1": jnp.array([5, 6]),
           "c2": jnp.array([7, 8]),
+          "c3": jnp.array([9, 10]),
+          "c4": jnp.array([11, 12]),
           "b1": jnp.array([True, False]),
           "b2": jnp.array([False, True]),
       })
@@ -594,18 +698,20 @@ class InternalFeatureSpecTest(parameterized.TestCase):
     internal_values = self.basic_mapping.convert_features({
         "n1": jnp.array([1, 2]),
         "n2": jnp.array([3, 4]),
-        "multidim_n3": jnp.array([[9, 10], [11, 12]]),
+        "multidim_n3": jnp.array([[13, 14], [15, 16]]),
         "c1": jnp.array([5, 6]),
         "c2": jnp.array([7, 8]),
+        "c3": jnp.array([9, 10]),
+        "c4": jnp.array([11, 12]),
         "b1": jnp.array([True, False]),
         "b2": jnp.array([False, True]),
     })
     np.testing.assert_array_equal(
         internal_values.numerical,
-        jnp.array([[9.0, 10.0, 1.0, 3.0], [11.0, 12.0, 2.0, 4.0]]),
+        jnp.array([[13.0, 14.0, 1.0, 3.0], [15.0, 16.0, 2.0, 4.0]]),
     )
     np.testing.assert_array_equal(
-        internal_values.categorical, jnp.array([[5, 7], [6, 8]])
+        internal_values.categorical, jnp.array([[5, 7, 9, 11], [6, 8, 10, 12]])
     )
     np.testing.assert_array_equal(
         internal_values.boolean, jnp.array([[True, False], [False, True]])
@@ -681,14 +787,17 @@ class InternalForestTest(parameterized.TestCase):
     )
     self.assertEqual(
         internal_forest.feature_spec.categorical,
-        [InternalFeatureItem(name="c1", dim=1)],
+        [
+            InternalFeatureItem(name="c1", dim=1),
+            InternalFeatureItem(name="c2", dim=1),
+        ],
     )
     self.assertEqual(internal_forest.feature_spec.boolean, [])
     self.assertEqual(internal_forest.feature_spec.inv_numerical, {1: 0, 3: 1})
-    self.assertEqual(internal_forest.feature_spec.inv_categorical, {2: 0})
+    self.assertEqual(internal_forest.feature_spec.inv_categorical, {2: 0, 4: 1})
     self.assertEqual(internal_forest.feature_spec.inv_boolean, {})
     self.assertEqual(
-        internal_forest.feature_spec.feature_names, {"f1", "f2", "c1"}
+        internal_forest.feature_spec.feature_names, {"f1", "f2", "c1", "c2"}
     )
 
     self.assertEqual(internal_forest.num_trees(), 2)
@@ -752,9 +861,9 @@ class InternalForestTest(parameterized.TestCase):
     self.assertEqual(internal_forest.max_depth, 3)
 
   def test_internal_forest_on_model(self):
-    columns = ["f1", "i1", "c1", "label_regress"]
+    columns = ["f1", "i1", "c1", "label_regress1"]
     model = specialized_learners.RandomForestLearner(
-        label="label_regress",
+        label="label_regress1",
         task=generic_learner.Task.REGRESSION,
         num_trees=10,
         max_depth=5,
@@ -785,31 +894,39 @@ class ToJaxTest(parameterized.TestCase):
       (
           "gbt_regression_num",
           ["f1", "f2"],
-          "label_regress",
+          "label_regress1",
           generic_learner.Task.REGRESSION,
           False,
           specialized_learners.GradientBoostedTreesLearner,
       ),
       (
           "gbt_regression_num_cat",
-          ["f1", "f2", "c1"],
-          "label_regress",
+          ["f1", "f2", "c1", "c2"],
+          "label_regress1",
           generic_learner.Task.REGRESSION,
           True,
           specialized_learners.GradientBoostedTreesLearner,
       ),
       (
           "gbt_class_binary_num_cat",
-          ["f1", "f2", "c1", "label_class_binary"],
-          "label_class_binary",
+          ["f1", "f2", "c1", "c2", "label_class_binary1"],
+          "label_class_binary1",
           generic_learner.Task.CLASSIFICATION,
           True,
           specialized_learners.GradientBoostedTreesLearner,
       ),
       (
-          "gbt_class_multi_num_cat",
-          ["f1", "f2", "c1", "label_class_multi"],
-          "label_class_multi",
+          "gbt_class_multi_num_cat1",
+          ["f1", "f2", "c1", "c2", "label_class_multi1"],
+          "label_class_multi1",
+          generic_learner.Task.CLASSIFICATION,
+          True,
+          specialized_learners.GradientBoostedTreesLearner,
+      ),
+      (
+          "gbt_class_multi_num_cat2",
+          ["f1", "f2", "c1", "c2", "label_class_multi2"],
+          "label_class_multi2",
           generic_learner.Task.CLASSIFICATION,
           True,
           specialized_learners.GradientBoostedTreesLearner,
@@ -817,7 +934,7 @@ class ToJaxTest(parameterized.TestCase):
       (
           "gbt_regression_num_multidim",
           ["f1", "multi_f1"],
-          "label_regress",
+          "label_regress1",
           generic_learner.Task.REGRESSION,
           False,
           specialized_learners.GradientBoostedTreesLearner,
@@ -825,7 +942,7 @@ class ToJaxTest(parameterized.TestCase):
       (
           "gbt_regression_num_oblique",
           ["f1", "f2", "f3", "f4"],
-          "label_regress",
+          "label_regress1",
           generic_learner.Task.REGRESSION,
           False,
           specialized_learners.GradientBoostedTreesLearner,